+++ test/replay/missing_replay_file.t
@@ -1,0 +1,5 @@
+  $ owi replay missing_replay_file.wat
+  owi: required option --replay-file is missing
+  Usage: owi replay [OPTION]… ARG
+  Try 'owi replay --help' or 'owi --help' for more information.
+  [124]